+++ CHANGELOG.md
@@ -5,8 +5,11 @@
 - React v18 is now required as peer dependency.
 - Removed `useMapElement` hook.
 - Removed `MapConsumer` component.
+- Removed `whenCreated` property from the `MapContainer` component.
+- Removed `onOpen` and `onClose` properties from `Popup` and `Tooltip` components.
 - Removed UMD distribution.
 
 ### Other changes
 
 - Added support for React v18 (concurrent mode).
+- Added support for `ref` in the `MapContainer` and `Pane` components.